EC2/airflow/requirements.txt
beautifulsoup4
dotenv
psycopg2-binary
requests
python-dateutil
lxml
google-generativeai